+++ src/renderer/src/features/dashboardSlice.js
@@ -6,19 +6,19 @@ import {
   REQUEST_MP_GET_EVENT_LOG_HISTORY,
   RESPONSE_RP_GET_EVENT_LOG_HISTORY
 } from '../../../main/utils/IPCEvents'
-import { requestGraphData, requestCustomGraphData } from '../components/dashboard/requestGraphData'
+import { requestGraphData } from '../components/dashboard/requestGraphData'
 
-export const showCustomTableData = (payload) => (dispatch) => {
-  dispatch(updateCustomTableData(payload))
-}
-export const showSyslogTableData = (payload) => (dispatch) => {
-  dispatch(updateSyslogTableData(payload))
-  dispatch(openDialog('syslogGraphTable'))
-}
-export const showTrapTableData = (payload) => (dispatch) => {
-  dispatch(updateTrapTableData(payload))
-  dispatch(openDialog('trapGraphTable'))
-}
+// export const showCustomTableDatashowCustomTableData = (payload) => (dispatch) => {
+//   dispatch(updateCustomTableData(payload))
+// }
+// export const showSyslogTableData = (payload) => (dispatch) => {
+//   dispatch(updateSyslogTableData(payload))
+//   dispatch(openDialog('syslogGraphTable'))
+// }
+// export const showTrapTableData = (payload) => (dispatch) => {
+//   dispatch(updateTrapTableData(payload))
+//   dispatch(openDialog('trapGraphTable'))
+// }
 export const requestHistoryData = (param) => (dispatch) => {
   window.electron.ipcRenderer.on(RESPONSE_RP_GET_EVENT_LOG_HISTORY, (event, arg) => {
     const { type, data } = arg
@@ -116,22 +116,36 @@ const dashboardSlice = createSlice({
         }
       }
     },
-    updateCustomGraph: (state, { payload }) => {
-      const { label, tableResult, lastUpdated, InformationData, CriticalData, WarningData } =
-        payload
-      // const { payload } = action
+    updateCustomGraph: (state, action) => {
+      const { payload } = action
       return {
         ...state,
         customGraphData: {
-          label: label,
-          InformationData: InformationData,
-          CriticalData: CriticalData,
-          tableData: tableResult,
-          WarningData: WarningData,
-          lastUpdated: lastUpdated
+          label: payload.label,
+          InformationData: payload.InformationData,
+          CriticalData: payload.CriticalData,
+          WarningData: payload.WarningData,
+          tableData: payload.tableResult,
+          lastUpdated: payload.lastUpdated
         }
       }
     },
+    // updateCustomGraph: (state, { payload }) => {
+    //   const { label, tableResult, lastUpdated, InformationData, CriticalData, WarningData } =
+    //     payload
+    //   // const { payload } = action
+    //   return {
+    //     ...state,
+    //     customGraphData: {
+    //       label: label,
+    //       InformationData: InformationData,
+    //       CriticalData: CriticalData,
+    //       tableData: tableResult,
+    //       WarningData: WarningData,
+    //       lastUpdated: lastUpdated
+    //     }
+    //   }
+    // },
 
     updateSyslogTableData: (state, { payload }) => {
       //  const { payload } = action
@@ -197,3 +211,80 @@ export const dashboardSelector = (state) => {
 }
 
 export default dashboardSlice
+
+export const requestCustomGraphData = (Items) => {
+  let label = []
+  let InformationData = []
+  let CriticalData = []
+  let WarningData = []
+  let tableResult = []
+  for (let index = 7; index > 0; index--) {
+    let le = ''
+    let ge = ''
+    let ledate = new Date()
+    ledate.setDate(ledate.getDate() - (index - 1 - 1))
+    let gedate = new Date()
+    gedate.setDate(gedate.getDate() - (index - 1))
+    le = `${ledate.getFullYear()}-${('00' + (ledate.getMonth() + 1)).slice(-2)}-${(
+      '00' + ledate.getDate()
+    ).slice(-2)}`
+    ge = `${gedate.getFullYear()}-${('00' + (gedate.getMonth() + 1)).slice(-2)}-${(
+      '00' + gedate.getDate()
+    ).slice(-2)}`
+    let result = Items.filter(function (item) {
+      return (
+        new Date(item.createAt).getTime() >= new Date(ge).getTime() &&
+        new Date(item.createAt).getTime() < new Date(le).getTime()
+      )
+    })
+    let resultInformation = Items.filter(function (item) {
+      return (
+        new Date(item.createAt).getTime() >= new Date(ge).getTime() &&
+        new Date(item.createAt).getTime() < new Date(le).getTime() &&
+        item.severity === 'Information'
+      )
+    })
+    let resultCritical = Items.filter(function (item) {
+      return (
+        new Date(item.createAt).getTime() >= new Date(ge).getTime() &&
+        new Date(item.createAt).getTime() < new Date(le).getTime() &&
+        item.severity === 'Critical'
+      )
+    })
+    let resultWarning = Items.filter(function (item) {
+      return (
+        new Date(item.createAt).getTime() >= new Date(ge).getTime() &&
+        new Date(item.createAt).getTime() < new Date(le).getTime() &&
+        item.severity === 'Warning'
+      )
+    })
+    let gelabel = `${('00' + (gedate.getMonth() + 1)).slice(-2)}/${('00' + gedate.getDate()).slice(
+      -2
+    )}`
+    label.push(gelabel)
+    InformationData.push(resultInformation.length)
+    CriticalData.push(resultCritical.length)
+    WarningData.push(resultWarning.length)
+    tableResult.push(result)
+  }
+  const date = new Date()
+  const lastUpdated =
+    'last update ' +
+    ('00' + date.getDate()).slice(-2) +
+    '/' +
+    ('00' + (date.getMonth() + 1)).slice(-2) +
+    '/' +
+    date.getFullYear() +
+    ' ' +
+    ('00' + date.getHours()).slice(-2) +
+    ':' +
+    ('00' + date.getMinutes()).slice(-2)
+  return {
+    label,
+    InformationData,
+    CriticalData,
+    WarningData,
+    lastUpdated,
+    tableResult
+  }
+}
